Make pubsub implementation classes package-private
