+++ modules/streams/src/main/scala/dev/profunktor/redis4cats/pubsub/internals/LivePubSubStats.scala
@@ -16,32 +16,38 @@
 
 package dev.profunktor.redis4cats
 package pubsub
+package internals
 
 import cats.effect._
-import cats.syntax.functor._
-import dev.profunktor.redis4cats.data.RedisChannel
+import cats.syntax.all._
+import dev.profunktor.redis4cats.data._
 import dev.profunktor.redis4cats.effect.{ JRFuture, RedisEc }
 import dev.profunktor.redis4cats.pubsub.data.Subscription
 import fs2.Stream
 import io.lettuce.core.pubsub.StatefulRedisPubSubConnection
 
-class Publisher[F[_]: ConcurrentEffect: ContextShift, K, V](
+import dev.profunktor.redis4cats.JavaConversions._
+
+private[pubsub] class LivePubSubStats[F[_]: Concurrent: ContextShift, K, V](
     pubConnection: StatefulRedisPubSubConnection[K, V],
     redisEc: RedisEc
-) extends PublishCommands[Stream[F, *], K, V] {
-
-  private[redis4cats] val pubSubStats: PubSubStats[Stream[F, *], K] = new LivePubSubStats(pubConnection, redisEc)
-
-  override def publish(channel: RedisChannel[K]): Stream[F, V] => Stream[F, Unit] =
-    _.evalMap(message => JRFuture(F.delay(pubConnection.async().publish(channel.underlying, message)))(redisEc).void)
+) extends PubSubStats[Stream[F, *], K] {
 
   override def pubSubChannels: Stream[F, List[K]] =
-    pubSubStats.pubSubChannels
+    Stream
+      .eval {
+        JRFuture(F.delay(pubConnection.async().pubsubChannels()))(redisEc)
+      }
+      .map(_.asScala.toList)
 
   override def pubSubSubscriptions(channel: RedisChannel[K]): Stream[F, Subscription[K]] =
-    pubSubStats.pubSubSubscriptions(channel)
+    pubSubSubscriptions(List(channel)).map(_.headOption).unNone
 
   override def pubSubSubscriptions(channels: List[RedisChannel[K]]): Stream[F, List[Subscription[K]]] =
-    pubSubStats.pubSubSubscriptions(channels)
+    Stream.eval {
+      JRFuture(F.delay(pubConnection.async().pubsubNumsub(channels.map(_.underlying): _*)))(redisEc).flatMap { kv =>
+        F.delay(kv.asScala.toList.map { case (k, n) => Subscription(RedisChannel[K](k), n) })
+      }
+    }
 
 }
