+++ modules/streams/src/main/scala/dev/profunktor/redis4cats/pubsub/internals/Publisher.scala
@@ -16,37 +16,33 @@
 
 package dev.profunktor.redis4cats
 package pubsub
+package internals
 
 import cats.effect._
-import cats.syntax.all._
-import dev.profunktor.redis4cats.data._
+import cats.syntax.functor._
+import dev.profunktor.redis4cats.data.RedisChannel
 import dev.profunktor.redis4cats.effect.{ JRFuture, RedisEc }
 import dev.profunktor.redis4cats.pubsub.data.Subscription
 import fs2.Stream
 import io.lettuce.core.pubsub.StatefulRedisPubSubConnection
 
-import dev.profunktor.redis4cats.JavaConversions._
-
-class LivePubSubStats[F[_]: Concurrent: ContextShift, K, V](
+private[pubsub] class Publisher[F[_]: ConcurrentEffect: ContextShift, K, V](
     pubConnection: StatefulRedisPubSubConnection[K, V],
     redisEc: RedisEc
-) extends PubSubStats[Stream[F, *], K] {
+) extends PublishCommands[Stream[F, *], K, V] {
+
+  private[redis4cats] val pubSubStats: PubSubStats[Stream[F, *], K] = new LivePubSubStats(pubConnection, redisEc)
+
+  override def publish(channel: RedisChannel[K]): Stream[F, V] => Stream[F, Unit] =
+    _.evalMap(message => JRFuture(F.delay(pubConnection.async().publish(channel.underlying, message)))(redisEc).void)
 
   override def pubSubChannels: Stream[F, List[K]] =
-    Stream
-      .eval {
-        JRFuture(F.delay(pubConnection.async().pubsubChannels()))(redisEc)
-      }
-      .map(_.asScala.toList)
+    pubSubStats.pubSubChannels
 
   override def pubSubSubscriptions(channel: RedisChannel[K]): Stream[F, Subscription[K]] =
-    pubSubSubscriptions(List(channel)).map(_.headOption).unNone
+    pubSubStats.pubSubSubscriptions(channel)
 
   override def pubSubSubscriptions(channels: List[RedisChannel[K]]): Stream[F, List[Subscription[K]]] =
-    Stream.eval {
-      JRFuture(F.delay(pubConnection.async().pubsubNumsub(channels.map(_.underlying): _*)))(redisEc).flatMap { kv =>
-        F.delay(kv.asScala.toList.map { case (k, n) => Subscription(RedisChannel[K](k), n) })
-      }
-    }
+    pubSubStats.pubSubSubscriptions(channels)
 
 }
